+++ ingest/operation_details.go
@@ -396,6 +396,12 @@ func (is *Session) operationDetails() map[string]interface{} {
 		details["dest_account_id"] = op.RedemptionRequest.Destination.Address()
 		details["reason"] = op.RedemptionRequest.CreatorDetails
 		details["reference"] = op.Reference
+	case xdr.OperationTypeCreateDataCreationRequest:
+	case xdr.OperationTypeCreateDataUpdateRequest:
+	case xdr.OperationTypeCreateDataRemoveRequest:
+	case xdr.OperationTypeCancelDataCreationRequest:
+	case xdr.OperationTypeCancelDataUpdateRequest:
+	case xdr.OperationTypeCancelDataRemoveRequest:
 	default:
 		panic(fmt.Errorf("Unknown operation type: %s", c.OperationType()))
 	}