+++ ingest/participants/main.go
@@ -212,6 +212,12 @@ func ForOperation(
 	case xdr.OperationTypeUpdateData:
 	case xdr.OperationTypeCreateRedemptionRequest:
 		// TODO add participant
+	case xdr.OperationTypeCreateDataCreationRequest:
+	case xdr.OperationTypeCreateDataUpdateRequest:
+	case xdr.OperationTypeCreateDataRemoveRequest:
+	case xdr.OperationTypeCancelDataCreationRequest:
+	case xdr.OperationTypeCancelDataUpdateRequest:
+	case xdr.OperationTypeCancelDataRemoveRequest:
 	default:
 		err = fmt.Errorf("unknown operation type: %s", op.Body.Type)
 	}